requirements.txt
os
subprocess
tempfile
IPython.core.magic
IPython.core.magic_arguments
